requirements.txt
asgiref==3.3.1
Django==3.1.6
django-bootstrap-pagination==1.7.1
django-filter==2.4.0
django-widget-tweaks==1.4.8
Pillow==8.1.0
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
